pyproject.toml
[project]
name = "ses-email"
version = "0.1.0"
description = "SES Email Sender with TUI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.24",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "textual>=1.0.0",
    "openpyxl>=3.1.0",
    "rich>=14.2.0",
]

[project.scripts]
ses-email = "tui.app:main"
